requirements.txt
absl-py==2.0.0
accelerate==0.21.0
addict==2.4.0
aenum==3.1.15
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.1
alembic==1.13.2
altair==5.2.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
apturl==0.5.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.1.0
av==11.0.0
basicsr==1.4.2
bcrypt==3.2.0
beautifulsoup4==4.12.2
blendmodes==2022
blinker==1.4
boltons==23.1.1
Brlapi==0.8.3
cachetools==5.3.2
certifi==2020.6.20
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.0.3
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip
cmake==3.28.1
colorama==0.4.4
coloredlogs==15.0.1
colorlog==6.8.2
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cssselect2==0.7.0
cupshelpers==1.0
cycler==0.12.1
Cython==3.0.6
dbus-python==1.2.18
defer==1.0.6
deprecation==2.1.0
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.0.6
docker-pycreds==0.4.0
duplicity==0.8.21
easydict==1.11
einops==0.4.1
exceptiongroup==1.2.0
facexlib==0.3.0
fastapi==0.94.0
fasteners==0.14.1
ffmpy==0.3.1
filelock==3.13.1
filterpy==1.4.5
flatbuffers==23.5.26
fonttools==4.46.0
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
future==0.18.2
fvcore==0.1.5.post20221221
gdown==4.7.1
gfpgan==1.3.8
gitdb==4.0.11
GitPython==3.1.32
google-auth==2.25.2
google-auth-oauthlib==1.2.0
grad-cam==1.5.2
gradio==3.41.2
gradio_client==0.5.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.12.0
httpcore==0.15.0
httplib2==0.20.2
httpx==0.24.1
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.3
ifnude==0.0.3
image==1.5.33
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.0
importlib-resources==6.1.1
inflection==0.5.1
insightface==0.7.3
iopath==0.1.9
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
jsonmerge==1.8.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
keyring==23.5.0
kiwisolver==1.4.5
kornia==0.6.7
language-selector==0.1
lark==1.1.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
lightning-utilities==0.10.0
lit==17.0.6
llvmlite==0.41.1
lmdb==1.4.1
lockfile==0.12.2
louis==3.20.0
lpips==0.1.4
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.0
mdurl==0.1.2
mediapipe==0.10.9
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
netifaces==0.11.0
networkx==3.2.1
numba==0.58.1
numexpr==2.8.8
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
omegaconf==2.2.3
onnx==1.15.0
onnxruntime==1.16.3
onnxruntime-gpu==1.16.3
open-clip-torch==2.20.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
optuna==3.6.1
orjson==3.9.10
packaging==23.2
pandas==2.1.4
paramiko==2.9.3
pexpect==4.8.0
piexif==1.1.3
Pillow==9.5.0
PIMS==0.6.0
platformdirs==4.1.0
portalocker==2.8.2
prettytable==3.9.0
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.13
pydub==0.25.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-multipart==0.0.6
pytorch-lightning==1.9.4
pytorch-triton-rocm==2.1.0
pytz==2022.1
PyWavelets==1.5.0
pyxdg==0.27
PyYAML==5.4.1
qudida==0.0.4
realesrgan==0.3.0
referencing==0.32.0
regex==2023.10.3
reportlab==3.6.8
requests==2.32.3
requests-oauthlib==1.3.1
resize-right==0.0.2
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.11.4
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.5.1
setproctitle==1.3.3
six==1.16.0
slicerator==1.1.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.0
ssh-import-id==5.11
starlette==0.26.1
svglib==1.5.1
sympy==1.12
systemd-python==234
tabulate==0.9.0
tb-nightly==2.16.0a20231217
tensorboard-data-server==0.7.2
termcolor==2.4.0
tf-keras-nightly==2.16.0.dev2023121710
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.9.2
tinycss2==1.2.1
tokenizers==0.13.3
tomesd==0.1.3
tomli==2.0.1
toolz==0.12.0
torch==2.3.1
torchdiffeq==0.2.3
torchmetrics==1.2.1
torchsde==0.2.5
torchvision==0.18.1
tqdm==4.66.1
trampoline==0.1.2
transformers==4.30.2
triton==2.3.1
ttach==0.0.3
typing_extensions==4.9.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ultralytics==8.0.227
unattended-upgrades==0.1
urllib3==2.2.2
usb-creator==0.3.7
uvicorn==0.15.0
wadllib==1.3.6
wandb==0.17.4
wcwidth==0.2.12
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
wsproto==1.0.0
xdg==5
xformers==0.0.23.post1
xkit==0.0.0
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==1.0.0